# app-template/openapi/types/createEventRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Event Management API
 * API for Event Management System
 * OpenAPI spec version: 1.0.0
 */
import type { CreateEventRequestVisibility } from './createEventRequestVisibility';
import type { CreateEventRequestRegistrationStatus } from './createEventRequestRegistrationStatus';
import type { CreateEventRequestStatus } from './createEventRequestStatus';

export interface CreateEventRequest {
  /** @maxLength 255 */
  name: string;
  startDate: Date;
  endDate: Date;
  /**
   * @maxLength 255
   * @nullable
   */
  venueName?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  description?: string | null;
  /**
   * @minimum 0
   * @nullable
   */
  capacity?: number | null;
  visibility?: CreateEventRequestVisibility;
  registrationStatus?: CreateEventRequestRegistrationStatus;
  status?: CreateEventRequestStatus;
}