# app-template/openapi/types/errorResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Event Management API
 * API for Event Management System
 * OpenAPI spec version: 1.0.0
 */
import type { ErrorResponseFieldErrorsItem } from './errorResponseFieldErrorsItem';

export interface ErrorResponse {
  code: string;
  message: string;
  details?: string;
  fieldErrors?: ErrorResponseFieldErrorsItem[];
}