# app-template/openapi/types/eventResponse.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Event Management API
 * API for Event Management System
 * OpenAPI spec version: 1.0.0
 */
import type { EventResponseStatus } from './eventResponseStatus';
import type { EventResponseVisibility } from './eventResponseVisibility';
import type { EventResponseRegistrationStatus } from './eventResponseRegistrationStatus';

export interface EventResponse {
  id: string;
  name: string;
  slug: string;
  startDate: Date;
  endDate: Date;
  /** @nullable */
  venueName?: string | null;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  capacity?: number | null;
  status: EventResponseStatus;
  visibility: EventResponseVisibility;
  registrationStatus: EventResponseRegistrationStatus;
  createdAt: Date;
  updatedAt: Date;
}