# app-template/openapi/types/paginationMeta.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Event Management API
 * API for Event Management System
 * OpenAPI spec version: 1.0.0
 */

export interface PaginationMeta {
  page: number;
  limit: number;
  totalPages: number;
  totalItems: number;
}