# app-template/openapi/types/registerUserRequest.ts
/**
 * Generated by orval v7.11.2 🍺
 * Do not edit manually.
 * Event Management API
 * API for Event Management System
 * OpenAPI spec version: 1.0.0
 */

export interface RegisterUserRequest {
  /** @maxLength 255 */
  email: string;
  /** @maxLength 100 */
  firstName: string;
  /** @maxLength 100 */
  lastName: string;
}